Use long instead of byte

+++ musig/src/main/java/com/chainx/musig/Mast.java
@@ -7,15 +7,15 @@ public class Mast {
         System.loadLibrary("musig_dll");
     }
 
-    public static native String generate_threshold_pubkey(String pubkeys, byte threshold);
+    public static native String generate_threshold_pubkey(String pubkeys, long threshold);
 
-    public static native String generate_control_block(String pubkeys, byte threshold, String sigAggPubkey);
+    public static native String generate_control_block(String pubkeys, long threshold, String sigAggPubkey);
 
-    public static String generateThresholdPubkey(String[] pubkeys, byte threshold) {
+    public static String generateThresholdPubkey(String[] pubkeys, long threshold) {
         return generate_threshold_pubkey(TextUtils.join("", pubkeys).toString(), threshold);
     }
 
-    public static String generateControlBlock(String[] pubkeys, byte threshold, String sigAggPubkey) {
+    public static String generateControlBlock(String[] pubkeys, long threshold, String sigAggPubkey) {
         return generate_control_block(TextUtils.join("", pubkeys).toString(), threshold, sigAggPubkey);
     }
